+++ packages/filesystem/src/node/filesystem-watcher-client.ts
@@ -14,74 +14,59 @@
  * SPDX-License-Identifier: EPL-2.0 OR GPL-2.0 WITH Classpath-exception-2.0
  ********************************************************************************/
 
-import * as path from 'path';
 import { injectable, inject } from 'inversify';
-import { JsonRpcProxyFactory, ILogger, ConnectionErrorHandler, DisposableCollection, Disposable } from '@theia/core';
-import { IPCConnectionProvider } from '@theia/core/lib/node/messaging';
-import { FileSystemWatcherServer, WatchOptions, FileSystemWatcherClient, ReconnectingFileSystemWatcherServer } from '../common/filesystem-watcher-protocol';
-import { NsfwOptions } from './nsfw-watcher/nsfw-options';
+import { FileSystemWatcherServer, WatchOptions, FileSystemWatcherClient, FileSystemWatcherService } from '../common/filesystem-watcher-protocol';
+import { FileSystemWatcherServiceDispatcher } from './filesystem-watcher-dispatcher';
 
 export const NSFW_WATCHER = 'nsfw-watcher';
 
+/**
+ * Wraps the NSFW singleton service for each frontend.
+ */
 @injectable()
 export class FileSystemWatcherServerClient implements FileSystemWatcherServer {
 
-    protected readonly proxyFactory = new JsonRpcProxyFactory<FileSystemWatcherServer>();
-    protected readonly remote = new ReconnectingFileSystemWatcherServer(this.proxyFactory.createProxy());
+    protected static clientIdSequence = 0;
 
-    protected readonly toDispose = new DisposableCollection();
+    /**
+     * Track allocated watcherIds to de-allocate on disposal.
+     */
+    protected watcherIds = new Set<number>();
 
-    constructor(
-        @inject(ILogger) protected readonly logger: ILogger,
-        @inject(IPCConnectionProvider) protected readonly ipcConnectionProvider: IPCConnectionProvider,
-        @inject(NsfwOptions) protected readonly nsfwOptions: NsfwOptions
-    ) {
-        this.remote.setClient({
-            onDidFilesChanged: e => {
-                if (this.client) {
-                    this.client.onDidFilesChanged(e);
-                }
-            },
-            onError: () => {
-                if (this.client) {
-                    this.client.onError();
-                }
-            }
-        });
-        this.toDispose.push(this.remote);
-        this.toDispose.push(this.listen());
+    /**
+     * @todo make this number precisely map to one specific frontend.
+     */
+    protected readonly clientId = FileSystemWatcherServerClient.clientIdSequence++;
+
+    @inject(FileSystemWatcherServiceDispatcher)
+    protected readonly watcherDispatcher: FileSystemWatcherServiceDispatcher;
+
+    @inject(FileSystemWatcherService)
+    protected readonly watcherService: FileSystemWatcherService;
+
+    async watchFileChanges(uri: string, options?: WatchOptions): Promise<number> {
+        const watcherId = await this.watcherService.watchFileChanges(this.clientId, uri, options);
+        this.watcherIds.add(watcherId);
+        return watcherId;
+    }
+
+    async unwatchFileChanges(watcherId: number): Promise<void> {
+        this.watcherIds.delete(watcherId);
+        return this.watcherService.unwatchFileChanges(watcherId);
+    }
+
+    setClient(client: FileSystemWatcherClient | undefined): void {
+        if (client !== undefined) {
+            this.watcherDispatcher.registerClient(this.clientId, client);
+        } else {
+            this.watcherDispatcher.unregisterClient(this.clientId);
+        }
     }
 
     dispose(): void {
-        this.toDispose.dispose();
+        this.setClient(undefined);
+        for (const watcherId of this.watcherIds) {
+            this.unwatchFileChanges(watcherId);
+        }
     }
-
-    watchFileChanges(uri: string, options?: WatchOptions): Promise<number> {
-        return this.remote.watchFileChanges(uri, options);
-    }
-
-    unwatchFileChanges(watcher: number): Promise<void> {
-        return this.remote.unwatchFileChanges(watcher);
-    }
-
-    protected client: FileSystemWatcherClient | undefined;
-    setClient(client: FileSystemWatcherClient | undefined): void {
-        this.client = client;
-    }
-
-    protected listen(): Disposable {
-        return this.ipcConnectionProvider.listen({
-            serverName: NSFW_WATCHER,
-            entryPoint: path.resolve(__dirname, NSFW_WATCHER),
-            errorHandler: new ConnectionErrorHandler({
-                serverName: NSFW_WATCHER,
-                logger: this.logger
-            }),
-            args: [
-                `--nsfwOptions=${JSON.stringify(this.nsfwOptions)}`
-            ],
-            env: process.env
-        }, connection => this.proxyFactory.listen(connection));
-    }
-
 }