+++ packages/filesystem/src/node/nsfw-watcher/nsfw-filesystem-watcher.spec.ts
@@ -20,7 +20,7 @@ import * as fs from 'fs-extra';
 import * as assert from 'assert';
 import URI from '@theia/core/lib/common/uri';
 import { FileUri } from '@theia/core/lib/node';
-import { NsfwFileSystemWatcherServer } from './nsfw-filesystem-watcher';
+import { NsfwFileSystemWatcherService } from './nsfw-filesystem-service';
 import { DidFilesChangedParams } from '../../common/filesystem-watcher-protocol';
 /* eslint-disable no-unused-expressions */
 
@@ -30,27 +30,26 @@ const track = temp.track();
 describe('nsfw-filesystem-watcher', function (): void {
 
     let root: URI;
-    let watcherServer: NsfwFileSystemWatcherServer;
+    let watcherService: NsfwFileSystemWatcherService;
     let watcherId: number;
 
     this.timeout(10000);
 
     beforeEach(async () => {
         root = FileUri.create(fs.realpathSync(temp.mkdirSync('node-fs-root')));
-        watcherServer = createNsfwFileSystemWatcherServer();
-        watcherId = await watcherServer.watchFileChanges(root.toString());
+        watcherService = createNsfwFileSystemWatcherService();
+        watcherId = await watcherService.watchFileChanges(0, root.toString());
         await sleep(2000);
     });
 
     afterEach(async () => {
         track.cleanupSync();
-        watcherServer.dispose();
+        watcherService.dispose();
     });
 
     it('Should receive file changes events from in the workspace by default.', async function (): Promise<void> {
         if (process.platform === 'win32') {
             this.skip();
-            return;
         }
         const actualUris = new Set<string>();
 
@@ -61,7 +60,7 @@ describe('nsfw-filesystem-watcher', function (): void {
             onError(): void {
             }
         };
-        watcherServer.setClient(watcherClient);
+        watcherService.setClient(watcherClient);
 
         const expectedUris = [
             root.resolve('foo').toString(),
@@ -87,7 +86,6 @@ describe('nsfw-filesystem-watcher', function (): void {
     it('Should not receive file changes events from in the workspace by default if unwatched', async function (): Promise<void> {
         if (process.platform === 'win32') {
             this.skip();
-            return;
         }
         const actualUris = new Set<string>();
 
@@ -98,10 +96,10 @@ describe('nsfw-filesystem-watcher', function (): void {
             onError(): void {
             }
         };
-        watcherServer.setClient(watcherClient);
+        watcherService.setClient(watcherClient);
 
         /* Unwatch root */
-        watcherServer.unwatchFileChanges(watcherId);
+        watcherService.unwatchFileChanges(watcherId);
 
         fs.mkdirSync(FileUri.fsPath(root.resolve('foo')));
         expect(fs.statSync(FileUri.fsPath(root.resolve('foo'))).isDirectory()).to.be.true;
@@ -118,8 +116,8 @@ describe('nsfw-filesystem-watcher', function (): void {
         assert.deepEqual(0, actualUris.size);
     });
 
-    function createNsfwFileSystemWatcherServer(): NsfwFileSystemWatcherServer {
-        return new NsfwFileSystemWatcherServer({
+    function createNsfwFileSystemWatcherService(): NsfwFileSystemWatcherService {
+        return new NsfwFileSystemWatcherService({
             verbose: true
         });
     }